Add project metatags
