+++ pages/collection/_short_name/index.vue
@@ -202,7 +202,7 @@ import sortBy from 'lodash/sortBy'
 import uniqBy from 'lodash/uniqBy'
 import 'vue-awesome/icons/star'
 import localConfig from '@/local.config'
-import { collectionMetaTags } from '@/mixins/collectionMetaTags'
+import { collectionMetaTags } from '@/mixins/metaTags'
 import { licenses } from '@/mixins/licenses'
 import { fetchCollectionByName } from '@/mixins/fetchCollectionByName'
 import { computeShareUrl } from '@/mixins/computeShareUrl'
