+++ pages/collection/_short_name/projects/_id/_presenter.vue/index.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-import { collectionMetaTags } from '@/mixins/collectionMetaTags'
+import { projectMetaTags } from '@/mixins/metaTags'
 import { fetchCollectionByName } from '@/mixins/fetchCollectionByName'
 import { notifications } from '@/mixins/notifications'
 import isEmpty from 'lodash/isEmpty'
@@ -27,7 +27,7 @@ export default {
     return layouts[params.presenter] || 'collection-tabs'
   },
 
-  mixins: [ notifications, fetchCollectionByName, collectionMetaTags ],
+  mixins: [ notifications, fetchCollectionByName, projectMetaTags ],
 
   data () {
     return {
